# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WitelonBank",
        "description": "Dokumentacja API dla systemu bankowego WitelonBank. Pamiętaj o autoryzacji (przycisk 'Authorize' na górze) dla chronionych endpointów.",
        "contact": {
            "email": "support@witelonbank.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Główny serwer API"
        }
    ],
    "paths": {
        "/api/admin/konta": {
            "get": {
                "tags": [
                    "Admin-Konta"
                ],
                "summary": "Pobiera listę wszystkich kont użytkowników",
                "operationId": "fa66a653a68479e8a5007cdb6f1615ea",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numer strony",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista kont",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/KontoResource"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Brak uprawnień"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/konta/{idKonta}": {
            "get": {
                "tags": [
                    "Admin-Konta"
                ],
                "summary": "Pobiera szczegóły konkretnego konta użytkownika",
                "operationId": "ff93861e73ea7694a05d0e59f558eb8c",
                "parameters": [
                    {
                        "name": "idKonta",
                        "in": "path",
                        "description": "ID konta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Szczegóły konta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KontoResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Konto nie znalezione"
                    },
                    "403": {
                        "description": "Brak uprawnień"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/konta/{idKonta}/block": {
            "patch": {
                "tags": [
                    "Admin-Konta"
                ],
                "summary": "Blokuje konto użytkownika",
                "operationId": "61ccd804a28ae3ef76e5d3dc7f65a421",
                "parameters": [
                    {
                        "name": "idKonta",
                        "in": "path",
                        "description": "ID konta do zablokowania",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Konto zablokowane",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KontoResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Konto nie znalezione"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/konta/{idKonta}/unblock": {
            "patch": {
                "tags": [
                    "Admin-Konta"
                ],
                "summary": "Odblokowuje konto użytkownika",
                "operationId": "133b8ca74d266379349c4a4d218ab854",
                "parameters": [
                    {
                        "name": "idKonta",
                        "in": "path",
                        "description": "ID konta do odblokowania",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Konto odblokowane",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KontoResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Konto nie znalezione"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/konta/{idKonta}/limit": {
            "patch": {
                "tags": [
                    "Admin-Konta"
                ],
                "summary": "Aktualizuje limit przelewu dla konta użytkownika",
                "operationId": "bdc67feda7ed9e467e5b48ebf1289497",
                "parameters": [
                    {
                        "name": "idKonta",
                        "in": "path",
                        "description": "ID konta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "limit_przelewu"
                                ],
                                "properties": {
                                    "limit_przelewu": {
                                        "description": "Nowy limit przelewu",
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Limit zaktualizowany",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KontoResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Konto nie znalezione"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorGeneral": {
                "title": "Błąd Ogólny",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Wystąpił błąd."
                    }
                },
                "type": "object"
            },
            "KontoResource": {
                "title": "Konto Resource",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorGeneral"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "numer_konta": {
                                "type": "string",
                                "example": "PL12345678901234567890123456"
                            },
                            "aktualne_saldo": {
                                "type": "number",
                                "format": "float",
                                "example": 1500.75
                            },
                            "waluta": {
                                "type": "string",
                                "example": "PLN"
                            },
                            "limit_przelewu_dzienny": {
                                "type": "number",
                                "format": "float",
                                "example": 1000,
                                "nullable": true
                            },
                            "czy_zablokowane": {
                                "type": "boolean",
                                "example": false
                            },
                            "utworzono": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2023-05-15T10:00:00Z"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Wprowadź token JWT z prefiksem Bearer (np. 'Bearer eyJhbGciOiJIUzI1NiI...). Token uzyskasz po zalogowaniu.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}